requirements.txt
chromadb
langchain
langchain-community
langchain-openai
pypdf2
tiktoken
streamlit
python-dotenv
pypdf
